requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
presidio-analyzer
presidio-anonymizer
tabulate
streamlit
